requirements.txt
aiofiles==0.8.0
anyio==3.6.1
asgiref==3.5.2
astroid==2.11.5
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
cleo==0.8.1
click==8.1.3
clikit==0.6.2
crashtest==0.3.1
cycler==0.11.0
dill==0.3.5.1
distlib==0.3.4
fastapi==0.78.0
filelock==3.7.0
fonttools==4.33.3
h11==0.13.0
html5lib==1.1
httptools==0.4.0
idna==3.3
importlib-metadata==4.11.4
isort==5.10.1
Jinja2==3.1.2
keyring==23.5.1
kiwisolver==1.4.2
lazy-object-proxy==1.7.1
lockfile==0.12.2
MarkupSafe==2.1.1
matplotlib==3.5.2
mccabe==0.7.0
msgpack==1.0.3
numpy==1.22.4
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
Pillow==9.1.1
pkginfo==1.8.2
platformdirs==2.5.2
poetry==1.1.13
poetry-core==1.0.8
ptyprocess==0.7.0
pydantic==1.9.1
pylev==1.4.0
pylint==2.13.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
shellingham==1.4.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.1
watchgod==0.8.2
webencodings==0.5.1
websockets==10.3
wrapt==1.14.1
zipp==3.8.0